Initial state after `rails new -d postgresql -T .`
It's a good habit to commit auto-generated code seperate from human
written code.

In the event generated code causes a failure, being able to
deterministicaly verify it was the generated code is indispensible.

By committing the immediate output you can more easily
triangulate if the upgraded library or generated code is causing the
error, or if *your* code is causing an error.