Pre-emptively remove turbolinks
Turbolinks keeps your applications front end (javascripts and css)
loaded over multiple page loads. Creating a slightly "faster" page
load. This is generally A-OK; but it means your javascript state may
persist longer than you expect.

On a greenfield project it's probably fine to keep it in; but since it's
primarily for "performance" and can introduce unsettling defects I tend
to remove it.