Add failing check for user account registration
A feature test interacts with the application to provide a "sanity
check" as a feature is built. Writing them helps us "get in our users head"
before we start writing any of the code to make it work. As developers,
we frequently prioritize the the code over the user. Feature tests are a
relatively small way to spend more time thinking about the user.

While the final HTML and messages may not have these values; we have an
idea of what the form will look like.

As we look at this test; we see it's three distinct parts:

1. Set-up
2. Actions
3. Expectations

We "set up" by visiting the home page and clicking the "create user"
link.

We take Action by filling in the fields of the create user form and
clicking the create user button.

We confirm our Expectations by checking:
* The page content for a "you did it!" message
* The database for a Users existance
* The resulting web address to see we've arrived at our expected
  destination.