Bring account creation test in line with devise defaults
Since I didn't remember what Devise does by default, I had to change my
test to match the form that it presents.

To get these I used my browsers debugger tools to find the `id`
attribute on each of the form fields.

While you *can* use the content of the label (I.e "Email"); this makes
your tests brittle. A designer or wordsmith may want to change the
content; if so, it's better not to have to change your tests. When you
introduce internationalization; you'll wind up needing to change quite a
few tests as well.

You'll notice I had to use the content for the "Sign up" button. This is
because devise doesn't provide an id for the form submit button! Ah
well.