Add flash message to application layout
Devise provides confirmation through rails 'flash' hash. The flash hash
is a "read-once" place to store messages for the user interface. It's
mostly used for feedback like "An email has been sent!" or "Your image
is being uploaded and processed and will appear shortly!" and what not.

This also let me bring the rest of the tests expectations in line with
Devise's defaults.