Roll up unrolled loop in call to flash partial.
An unrolled loop is when you have several lines of code that are the
same except for a given value. In this case, we have an unrolled loop
of two: :notice and :alert.

By converting these to an array and looping over them, we completely
remove the duplication and add a small bit of complexity.