Add nickname to registration form
Rails has the notion of "strong parameters" which help protect your
models from exploitation by an unsavory character. This means we have
to tell devise that we *want* peopl to be able to provide their nick
name on signup. We do this in our ApplicationController using a before
filter restricted to devise_controllers

We also add the users nickname to the welcome screen. This allows a user
to visually confirm that they are logged in as the correct person and
inspire them to logout when they leave.