User may upload file
This implementation relies on local file systems, which won't work when
we push to Heroku.

A future commit will make this work when uploading to AWS.

Of important note is the use of "strong parameters" here. This is the
`safe_picture_params` method in the PicturesController. This ensures
that only fields we want users to change may be changed.