Extract shared error renderer
Some people would put this in a helper. I, personally, prefer to keep as
much of my HTML as possible in partials as that means they're super easy
for a designer to work with. Yes it's a *touch* more verbose; but
keeping the html in the HTML is "worth it" in my opinion