Enable S3 Uploads
Uploading our images to S3 is the only way we can make things work on
Heroku.

This means you'll now need to have an AWS account and an S3 bucket to
upload images in development. We're using `dotenv-rails` to make loading
the ENV variables super easy.

Everything else is setting up your initializer to have the correct buckt
name and access credentials to upload your files. In the test
environment we want to just upload files locally; as uploading to S3 is
quite slow.