User may see their pictures
This commit introduces the concept of nested resources. Nested resources
are a nice way to denote ownership. By default, nested resources just
use the top level controller. Because I expect the pictures index to
wind up showing all the public pictures, I decided to make a
controller explicitely for the users pictures.

You *could* put the new/create methods in the users pictures controller
as well, and it wouldn't hurt anything. Since pictures aren't likely to
be created for anyone *except* the currently logged in user, I see it as
reasonable to leave it within the pictures controller.

We also get to see the new html5 `figure` and `figcaption` tags!