Picture uses a uuid for its id
UUIDs are nearly globally unique strings. By using them for the id for
our pictures, it makes it almost impossible for a malicious user to
guess the URL used to view a picture.

This lays the groundwork for allowing a guest to view a picture shared
with them by the pictures owner